Fix ax203 camlib compilation warnings.
* ax203.c:
  * Adds missing .pp_64k initialization values to 'ax203_eeprom_info' array.
  * ax203_decode_image(...): Add some 'unsigned int' casts to image width/height comparisons in 'for' loops. These changes do not alter current code behavior.
  * ax203_encode_image(...): The 'x' and 'y' are only used in two basic 'for' loops where by signed-unsigned comparison they are treated as unsigned anyways so change the type to unsigned. Also add one 'unsigned int' cast to suppress compiler warning.
* ax203_compress_jpeg.c:
  * Add couple 'unsigned long' casts to suppress compiler warning about signed-unsigned compare. These changes do not alter current code behavior.

+++ camlibs/ax203/ax203_compress_jpeg.c
@@ -321,12 +321,12 @@ ax206_compress_jpeg(Camera *camera, int **in, uint8_t *outbuf, int out_size,
 			jpeg_finish_compress (&cinfo);
 
 			stop = 0;
-			for (i = 2; i < buf_size && !stop; i += size) {
+			for (i = 2; (unsigned long)i < buf_size && !stop; i += size) {
 				stop = buf[i] == 0xff && buf[i + 1] == 0xda;
 				i += 2;
 				size = (buf[i] << 8) | buf[i + 1];
 			}
-			if (i >= buf_size) {
+			if ((unsigned long)i >= buf_size) {
 				gp_log (GP_LOG_ERROR, "ax203",
 					"missing in ff da marker?");
 				return GP_ERROR_CORRUPTED_DATA;
